fix: copy, subs, tracking
